+++ ndk-build2/src/error.rs
@@ -48,7 +48,7 @@ pub enum NdkError {
     #[error("Invalid semver")]
     InvalidSemver,
     #[error("Command `{}` had a non-zero exit code.", format!("{:?}", .0).replace('"', ""))]
-    CmdFailed(Command),
+    CmdFailed(Box<Command>),
     #[error(transparent)]
     Serialize(#[from] quick_xml::SeError),
     #[error("String `{1}` is not a UID")]