+++ ndk-examples/src/jni_audio.rs
@@ -1,6 +1,9 @@
 use {
     android_logger::{Config, init_once},
-    jni::objects::{JIntArray, JObject, JObjectArray},
+    jni::{
+        ScopeToken, jni_sig, jni_str,
+        objects::{JIntArray, JObject, JObjectArray, JString},
+    },
     log::LevelFilter,
     winit::platform::android::activity::AndroidApp,
 };
@@ -15,24 +18,26 @@ const GET_DEVICES_OUTPUTS: jni::sys::jint = 2;
 
 fn enumerate_audio_devices() -> Result<(), Box<dyn std::error::Error>> {
     // Create a VM for executing Java calls
-    // 创建用于执行 Java 调用的 VM
     let ctx = ndk_context::android_context();
-    let vm = unsafe { jni::JavaVM::from_raw(ctx.vm().cast()) }?;
-    let context = unsafe { JObject::from_raw(ctx.context().cast()) };
-    let mut env = vm.attach_current_thread()?;
+    let vm = unsafe { jni::JavaVM::from_raw(ctx.vm().cast()) };
+    let mut scope = ScopeToken::default();
+    let mut env = unsafe { vm.attach_current_thread_guard(Default::default, &mut scope)? };
+    let env = env.borrow_env_mut();
+    let context = unsafe { JObject::from_raw(env, ctx.context().cast()) };
 
     // Query the global Audio Service
-    // 查询全局音频服务
-    let class_ctxt = env.find_class("android/content/Context")?;
-    let audio_service = env.get_static_field(class_ctxt, "AUDIO_SERVICE", "Ljava/lang/String;")?;
+    let class_ctx = env.find_class(jni_str!("android/content/Context"))?;
+    let audio_service = env.get_static_field(
+        class_ctx,
+        jni_str!("AUDIO_SERVICE"),
+        jni_sig!(java.lang.String),
+    )?;
 
     let audio_manager = env
         .call_method(
             context,
-            "getSystemService",
-            // JNI type signature needs to be derived from the Java API
-            // (ArgTys)ResultTy
-            "(Ljava/lang/String;)Ljava/lang/Object;",
+            jni_str!("getSystemService"),
+            jni_sig!((java.lang.String) -> java.lang.Object),
             &[(&audio_service).into()],
         )?
         .l()?;
@@ -41,42 +46,58 @@ fn enumerate_audio_devices() -> Result<(), Box<dyn std::error::Error>> {
     // 枚举输出设备
     let devices = env.call_method(
         audio_manager,
-        "getDevices",
-        "(I)[Landroid/media/AudioDeviceInfo;",
+        jni_str!("getDevices"),
+        jni_sig!((jint) -> [android.media.AudioDeviceInfo]),
         &[GET_DEVICES_OUTPUTS.into()],
     )?;
 
     println!("-- Output Audio Devices --");
 
-    let device_array = unsafe { JObjectArray::from_raw(devices.l()?.into_raw()) };
-    let len = env.get_array_length(&device_array)?;
+    let device_array = unsafe { JObjectArray::<JObject>::from_raw(env, *devices.l()?) };
+    let len = device_array.len(env)? as usize;
     for i in 0..len {
-        let device = env.get_object_array_element(&device_array, i)?;
+        let device = device_array.get_element(env, i)?;
 
         // Collect device information
-        // 收集设备信息
         // See https://developer.android.com/reference/android/media/AudioDeviceInfo
-        let product_name: String = {
-            let name =
-                env.call_method(&device, "getProductName", "()Ljava/lang/CharSequence;", &[])?;
-            let name = env.call_method(name.l()?, "toString", "()Ljava/lang/String;", &[])?;
-            env.get_string((&name.l()?).into())?.into()
+        let product_name = {
+            let name = env.call_method(
+                &device,
+                jni_str!("getProductName"),
+                jni_sig!(() -> java.lang.CharSequence),
+                &[],
+            )?;
+            let name = env
+                .call_method(
+                    name.l()?,
+                    jni_str!("toString"),
+                    jni_sig!(() -> java.lang.String),
+                    &[],
+                )?
+                .l()?;
+            unsafe { JString::from_raw(env, *name) }.to_string()
         };
-        let id = env.call_method(&device, "getId", "()I", &[])?.i()?;
-        let ty = env.call_method(&device, "getType", "()I", &[])?.i()?;
+        let id = env
+            .call_method(&device, jni_str!("getId"), jni_sig!(() -> jint), &[])?
+            .i()?;
+        let ty = env
+            .call_method(&device, jni_str!("getType"), jni_sig!(() -> jint), &[])?
+            .i()?;
 
         let sample_rates = {
             let sample_array = unsafe {
-                JIntArray::from_raw(
-                    env.call_method(&device, "getSampleRates", "()[I", &[])?
-                        .l()?
-                        .into_raw(),
-                )
+                let arr = env.call_method(
+                    &device,
+                    jni_str!("getSampleRates"),
+                    jni_sig!(() -> [jint]),
+                    &[],
+                )?;
+                JIntArray::from_raw(env, *arr.l()?)
             };
-            let len = env.get_array_length(&sample_array)?;
+            let len = sample_array.len(env)?;
 
             let mut sample_rates = vec![0; len as usize];
-            env.get_int_array_region(sample_array, 0, &mut sample_rates)?;
+            sample_array.get_region(env, 0, &mut sample_rates)?;
             sample_rates
         };